Move component NoScript from 'components' to 'components/_shared'
