+++ layouts/default.vue
@@ -36,9 +36,9 @@ import Vue from 'vue'
 import { LinkPropertyHref, MetaInfo } from 'vue-meta'
 import ScaleLoader from 'vue-spinner/src/ScaleLoader.vue'
 
+import NoScript from '@/components/_shared/NoScript.vue'
 import SideNavigation from '@/components/_shared/SideNavigation.vue'
 import DevelopmentModeMark from '@/components/DevelopmentModeMark.vue'
-import NoScript from '@/components/NoScript.vue'
 import Data from '@/data/data.json'
 import { convertDateToSimpleFormat } from '@/utils/formatDate'
 import { getLinksLanguageAlternative } from '@/utils/i18nUtils'
